% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{gnaf_remove_custom}
\alias{gnaf_remove_custom}
\title{Remove custom addresses by PID}
\usage{
gnaf_remove_custom(con, pids)
}
\arguments{
\item{con}{DBI connection.}

\item{pids}{Character vector of \code{address_detail_pid} values to remove.}
}
\value{
Invisibly, the number of rows deleted.
}
\description{
Remove custom addresses by PID
}
