% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gnaf_status}
\alias{gnaf_status}
\title{Report row counts for gnafr tables}
\usage{
gnaf_status(con)
}
\arguments{
\item{con}{DBI connection.}
}
\value{
A data.table with table name and row count.
}
\description{
Report row counts for gnafr tables
}
