% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{subset_shapefile}
\alias{subset_shapefile}
\title{Subset an `sf` object by a column value}
\usage{
subset_shapefile(sf_obj, var, values, invert = FALSE)
}
\arguments{
\item{sf_obj}{An `sf` polygon object.}

\item{var}{Character name of the column to filter on.}

\item{values}{Value or vector of values to keep (uses `%in%`).}

\item{invert}{If `TRUE`, keep rows not matching `values`.}
}
\value{
Subsetted `sf` object.
}
\description{
Subset an `sf` object by a column value
}
