% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gnaf_connect}
\alias{gnaf_connect}
\title{Connect to a gnafr DuckDB database}
\usage{
gnaf_connect(path, read_only = FALSE)
}
\arguments{
\item{path}{Path to the DuckDB file. Pass ":memory:" for an in-memory DB.}

\item{read_only}{Open in read-only mode.}
}
\value{
A DBI connection object.
}
\description{
Connect to a gnafr DuckDB database
}
